requirements.txt
juntagrico~=2.0.0
gunicorn~=25.1.0
psycopg>=3.1.8
django-modeltranslation
